Backend/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
pymongo
python-dotenv
email-validator
dnspython